requirements.txt
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
numba>=0.55.0
tqdm>=4.62.0
PyYAML>=6.0
jupyter>=1.0.0